Moved Admin Page and Basic Page to routes directory
