+++ FrontEnd/src/routes/AdminPage/AutoApproveDelay/AutoApproveDelay.module.css
