+++ FrontEnd/src/routes/AdminPage/DetailView/DeleteModal.jsx
