+++ FrontEnd/src/routes/AdminPage/DetailView/DeleteModal.module.css
