+++ FrontEnd/src/routes/AdminPage/DetailView/ProfileDetail.jsx
