+++ FrontEnd/src/routes/AdminPage/DetailView/ProfileDetail.module.css
