+++ FrontEnd/src/routes/AdminPage/DetailView/UserDetail.jsx
