+++ FrontEnd/src/routes/AdminPage/DetailView/UserDetail.module.css
