+++ FrontEnd/src/routes/AdminPage/MainPage/Text.js
