+++ FrontEnd/src/routes/BasicPage/SignUp/ResendActivationPage.jsx
